+++ screens/LoadingScreen.tsx
@@ -1,6 +1,6 @@
 import * as FileSystem from 'expo-file-system'
 import LottieView from 'lottie-react-native'
-import React from 'react'
+import React, { FC, ReactElement } from 'react'
 import {
   Dimensions,
   StyleSheet,
@@ -8,114 +8,68 @@ import {
   TouchableOpacity,
   View,
 } from 'react-native'
-import { connect } from 'react-redux'
 import Icon from '../assets/fonts/icon_font_config'
 import { scaleMultiplier } from '../constants'
+import { selector, useAppDispatch } from '../hooks'
 import { changeActiveGroup } from '../redux/actions/activeGroupActions'
+import { deleteLanguageData } from '../redux/actions/databaseActions'
+import { deleteGroup } from '../redux/actions/groupsActions'
+import { setIsInstallingLanguageInstance } from '../redux/actions/isInstallingLanguageInstanceActions'
 import {
-  deleteLanguageData,
   setHasFetchedLanguageData,
   setHasOnboarded,
   setLanguageCoreFilesDownloadProgress,
   setTotalLanguageCoreFilesToDownload,
-} from '../redux/actions/databaseActions'
-import { deleteGroup } from '../redux/actions/groupsActions'
-import { setIsInstallingLanguageInstance } from '../redux/actions/isInstallingLanguageInstanceActions'
+} from '../redux/actions/languageInstallationActions'
 import { activeGroupSelector } from '../redux/reducers/activeGroup'
 import { colors } from '../styles/colors'
 import { type } from '../styles/typography'
 import { getTranslations } from '../translations/translationsConfig'
 
-function mapStateToProps(state) {
-  var activeGroup = state.activeGroup ? activeGroupSelector(state) : null
-  var translations = state.activeGroup
-    ? getTranslations(activeGroupSelector(state).language)
-    : null
-  return {
-    languageCoreFilesDownloadProgress:
-      state.database.languageCoreFilesDownloadProgress,
-    totalLanguageCoreFilesToDownload:
-      state.database.totalLanguageCoreFilesToDownload,
-    hasInstalledFirstLanguageInstance:
-      state.database.hasInstalledFirstLanguageInstance,
-    storedDownloads: state.storedDownloads,
-    hasFetchedLanguageData: state.database.hasFetchedLanguageData,
-    actingLanguageID: state.database.actingLanguageID,
-    activeGroup: activeGroup,
-    groups: state.groups,
-    recentActiveGroup: state.database.recentActiveGroup,
-    isDark: state.settings.isDarkModeEnabled,
-    t: translations,
-    isConnected: state.network.isConnected,
-  }
-}
-
-function mapDispatchToProps(dispatch) {
-  return {
-    setIsInstallingLanguageInstance: (status) => {
-      dispatch(setIsInstallingLanguageInstance(status))
-    },
-    setHasOnboarded: (status) => {
-      dispatch(setHasOnboarded(status))
-    },
-    setTotalLanguageCoreFilesToDownload: (totalLanguageCoreFilesToDownload) => {
-      dispatch(
-        setTotalLanguageCoreFilesToDownload(totalLanguageCoreFilesToDownload)
-      )
-    },
-    setLanguageCoreFilesDownloadProgress: (progress) => {
-      dispatch(setLanguageCoreFilesDownloadProgress(progress))
-    },
-    setHasFetchedLanguageData: (hasFetchedLanguageData) => {
-      dispatch(setHasFetchedLanguageData(hasFetchedLanguageData))
-    },
-    deleteLanguageData: (language) => {
-      dispatch(deleteLanguageData(language))
-    },
-    deleteGroup: (groupName) => dispatch(deleteGroup(groupName)),
-    changeActiveGroup: (groupName) => dispatch(changeActiveGroup(groupName)),
-  }
-}
+interface Props {}
 
 /**
  * A screen that shows the download progress of a language's core files. The user sits on this screen if they finish onboarding before the downloads are done.
  */
-const LoadingScreen = ({
-  // Props passed from navigation.
-  navigation,
-  // Props passed from redux.
-  languageCoreFilesDownloadProgress,
-  totalLanguageCoreFilesToDownload,
-  hasInstalledFirstLanguageInstance,
-  storedDownloads,
-  actingLanguageID,
-  activeGroup,
-  groups,
-  recentActiveGroup,
-  hasFetchedLanguageData,
-  isDark,
-  t,
-  isConnected,
-  setIsInstallingLanguageInstance,
-  setHasOnboarded,
-  setTotalLanguageCoreFilesToDownload,
-  setLanguageCoreFilesDownloadProgress,
-  setHasFetchedLanguageData,
-  deleteLanguageData,
-  deleteGroup,
-  changeActiveGroup,
-}) => {
+const LoadingScreen: FC<Props> = ({ navigation }): ReactElement => {
+  const isDark = selector((state) => state.settings.isDarkModeEnabled)
+  const activeGroup = selector((state) => activeGroupSelector(state))
+  const t = getTranslations(activeGroup.language)
+  const isConnected = selector((state) => state.network.isConnected)
+  const languageCoreFilesDownloadProgress = selector(
+    (state) => state.languageInstallation.languageCoreFilesDownloadProgress
+  )
+  const totalLanguageCoreFilesToDownload = selector(
+    (state) => state.languageInstallation.totalLanguageCoreFilesToDownload
+  )
+  const hasInstalledFirstLanguageInstance = selector(
+    (state) => state.languageInstallation.hasInstalledFirstLanguageInstance
+  )
+  const storedDownloads = selector((state) => state.storedDownloads)
+  const hasFetchedLanguageData = selector(
+    (state) => state.languageInstallation.hasFetchedLanguageData
+  )
+  const actingLanguageID = selector(
+    (state) => state.languageInstallation.actingLanguageID
+  )
+  const groups = selector((state) => state.groups)
+  const recentActiveGroup = selector(
+    (state) => state.languageInstallation.recentActiveGroup
+  )
+
+  const dispatch = useAppDispatch()
+
   /** Cancels the language core files downloads, does a few cleanup actions, and sends the user back to the language instance install screen. */
   const cancelDownloads = () => {
     // Set the core files download progress to 0.
-    setLanguageCoreFilesDownloadProgress(0)
+    dispatch(setLanguageCoreFilesDownloadProgress(0))
 
     // Set this to 1 to avoid strange divide by zero errors.
-    setTotalLanguageCoreFilesToDownload(1)
+    dispatch(setTotalLanguageCoreFilesToDownload(1))
 
     // Set the isInstalling and hasFetched redux states to false since we're no longer installing.
-    setIsInstallingLanguageInstance(false)
-    setHasFetchedLanguageData(false)
+    dispatch(setIsInstallingLanguageInstance(false))
+    dispatch(setHasFetchedLanguageData(false))
 
     // Cancel all of the downloads.
     storedDownloads.forEach((download) =>
@@ -124,38 +78,39 @@ const LoadingScreen = ({
 
     // If we're adding a language for the first time, set hasOnboarded to false so that the user must go through onboarding again and reset them back to the language instance install screen.
     if (!hasInstalledFirstLanguageInstance) {
-      setHasOnboarded(false)
+      dispatch(setHasOnboarded(false))
       navigation.reset({
         index: 0,
-        routes: [{ name: 'InitialLanguageInstanceInstall' }],
+        routes: [{ name: 'InitialLanguageSelect' }],
       })
     }
 
     // If this is a subsequent language instance install, if we cancel, we need to switch back to our most recent active group since we're going to delete all the groups for the language we cancelled the installation of.
-    if (recentActiveGroup) changeActiveGroup(recentActiveGroup)
+    if (recentActiveGroup) dispatch(changeActiveGroup(recentActiveGroup))
     // If this is the initial language instance install, set active group back to null since there should be no active group before you install your first language instance.
-    else changeActiveGroup(null)
+    else dispatch(changeActiveGroup(''))
 
     // Delete any groups from the cancelled language.
     groups.forEach((group) => {
       if (group.language === actingLanguageID) {
-        deleteGroup(group.name)
+        dispatch(deleteGroup(group.name))
       }
     })
 
     // Delete the language data of the cancelled language.
-    deleteLanguageData(actingLanguageID)
+    if (actingLanguageID) dispatch(deleteLanguageData(actingLanguageID))
 
-    // Delete any files that have alreayd finished downloading for the cancelled language.
-    FileSystem.readDirectoryAsync(FileSystem.documentDirectory).then(
-      (contents) => {
-        for (const item of contents) {
-          if (item.slice(0, 2) === actingLanguageID) {
-            FileSystem.deleteAsync(FileSystem.documentDirectory + item)
+    // Delete any files that have already finished downloading for the cancelled language.
+    if (FileSystem.documentDirectory)
+      FileSystem.readDirectoryAsync(FileSystem.documentDirectory).then(
+        (contents) => {
+          for (const item of contents) {
+            if (item.slice(0, 2) === actingLanguageID) {
+              FileSystem.deleteAsync(FileSystem.documentDirectory + item)
+            }
           }
         }
-      }
-    )
+      )
   }
 
   return (
@@ -292,4 +247,4 @@ const styles = StyleSheet.create({
   },
 })
 
-export default connect(mapStateToProps, mapDispatchToProps)(LoadingScreen)
+export default LoadingScreen
